Decomposes tags service into `tagProvider` and `tagStore` services
This is in preparation for simpler customization of `tagProvider`
services beyond the current cache implemented in localStorage.

- The tag service now injects a tagProvider and tagStore
- The localStorage implementation has moved to local-tags.js
- local-tags.js is now injected back as *both* tagProvider and tagStore.

(Preserves existing functionality in client)
