+++ src/sidebar/services/tags.js
@@ -5,17 +5,19 @@
  * @property {number} updated - The timestamp when this tag was last used.
  */
 
+
 /**
- * Service for fetching tag suggestions and storing data to generate them.
+ * Service for fetching tag suggestions from a tagProvider service
+ * and storing data for future suggestions.
  *
- * The `tags` service stores metadata about recently used tags to local storage
- * and provides a `filter` method to fetch tags matching a query, ranked based
- * on frequency of usage.
+ * The injected `tagProvider` service needs to provide `filter` method to
+ * fetch tag suggestions matching a query and optional context object.
+ *
+ * The injected `tagStore` service needs to provide a `store`
+ * method to store entered tags.
  */
 // @inject
-export default function tags(localStorage) {
-  const TAGS_LIST_KEY = 'hypothesis.user.tags.list';
-  const TAGS_MAP_KEY = 'hypothesis.user.tags.map';
+export default function tags(tagProvider, tagStore) {
 
   /**
    * Return a list of tag suggestions matching `query`.
@@ -24,27 +26,9 @@ export default function tags(localStorage) {
    * @param {number|null} limit - Optional limit of the results.
    * @return {Tag[]} List of matching tags
    */
-  function filter(query, limit = null) {
-    const savedTags = localStorage.getObject(TAGS_LIST_KEY) || [];
-    let resultCount = 0;
-    // query will match tag if:
-    // * tag starts with query (e.g. tag "banana" matches query "ban"), OR
-    // * any word in the tag starts with query
-    //   (e.g. tag "pink banana" matches query "ban"), OR
-    // * tag has substring query occurring after a non-word character
-    //   (e.g. tag "pink!banana" matches query "ban")
-    let regex = new RegExp('(\\W|\\b)' + query, 'i');
-    return savedTags.filter(tag => {
-      if (tag.match(regex)) {
-        if (limit === null || resultCount < limit) {
-          // limit allows a subset of the results
-          // See https://github.com/hypothesis/client/issues/1606
-          ++resultCount;
-          return true;
-        }
-      }
-      return false;
-    });
+  function filter(query,
+                  limit = null) {
+    return tagProvider.filter(query, limit);
   }
 
   /**
@@ -54,30 +38,7 @@ export default function tags(localStorage) {
    * @param {Tag[]} tags - List of tags.
    */
   function store(tags) {
-    // Update the stored (tag, frequency) map.
-    const savedTags = localStorage.getObject(TAGS_MAP_KEY) || {};
-    tags.forEach(tag => {
-      if (savedTags[tag.text]) {
-        savedTags[tag.text].count += 1;
-        savedTags[tag.text].updated = Date.now();
-      } else {
-        savedTags[tag.text] = {
-          text: tag.text,
-          count: 1,
-          updated: Date.now(),
-        };
-      }
-    });
-    localStorage.setObject(TAGS_MAP_KEY, savedTags);
-
-    // Sort tag suggestions by frequency.
-    const tagsList = Object.keys(savedTags).sort((t1, t2) => {
-      if (savedTags[t1].count !== savedTags[t2].count) {
-        return savedTags[t2].count - savedTags[t1].count;
-      }
-      return t1.localeCompare(t2);
-    });
-    localStorage.setObject(TAGS_LIST_KEY, tagsList);
+    return tagStore.store(tags);
   }
 
   return {
